test-app-backend/requirements.txt
flask
gunicorn
sqlalchemy
python-dotenv
alembic
psycopg2-binary
flask-jwt-extended